Refactor clasp push
